avangard/requirements.txt
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
black==22.3.0
click==8.1.3
colorama==0.4.4
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.5
django-admin-rangefilter==0.8.4
django-ajax-selects==2.2.0
django-import-export==2.8.0
django-object-actions==4.0.0
et-xmlfile==1.1.0
executing==0.8.3
ipdb==0.13.9
ipython==8.4.0
jedi==0.18.1
MarkupPy==1.14
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
odfpy==1.4.1
openpyxl==3.0.10
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.29
pure-eval==0.2.2
Pygments==2.12.0
python-dotenv==0.20.0
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
tablib==3.2.1
toml==0.10.2
tomli==2.0.1
traitlets==5.2.1.post0
tzdata==2022.1
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
